refact

+++ sources/Device/src/libs/USBD/usbd_cdc_interface.cpp
@@ -121,7 +121,7 @@ static int8_t CDC_Itf_Control (uint8 cmd, uint8* pbuf, uint16)
 
 static int8_t CDC_Itf_Receive(uint8 *buffer, uint *length) //-V2009 //-V2558
 {
-    Message::SCPI::Data(buffer, *length).Transmit();
+    Message::SCPI::Data(buffer, static_cast<int>(*length)).Transmit();
 
     USBD_CDC_ReceivePacket(&DVCP::handleUSBD);
