+++ sources/Panel/src/Hardware/VCP_p.cpp
@@ -10,5 +10,5 @@ void PVCP::SendStringAsynch(const char *data)
 {
     char *mes = const_cast<char *>(data);
 
-    Message::SCPI::Data(reinterpret_cast<uint8 *>(mes), std::strlen(data)).Transmit();
+    Message::SCPI::Data(reinterpret_cast<uint8 *>(mes), static_cast<int>(std::strlen(data))).Transmit();
 }
